Splitted the ProcessMonitor to have a cleaner separation of windows-specific and corss-platform code
